tree-master/requirements.txt
flask~=1.1.2
pillow~=8.4.0
numpy~=1.24.3
tensorflow~=2.10.0